requirements.txt
Babel==2.4.0
cairocffi==0.8.0
CairoSVG==1.0.22
cffi==1.10.0
cssselect==1.0.1
Django==1.10.7
django-appconf==1.0.2
django-compressor==2.1.1
django-extra-views==0.6.4
django-haystack==2.6.0
django-lockdown==1.4.2
django-oscar==1.4
django-tables2==1.4.2
django-treebeard==4.1.0
django-widget-tweaks==1.4.1
factory-boy==2.8.1
Faker==0.7.11
funcsigs==1.0.2
html5lib==0.999999999
ipaddress==1.0.18
lxml==3.7.1
mock==2.0.0
olefile==0.44
pbr==2.1.0
phonenumbers==8.4.0
Pillow==4.1.0
psycopg2==2.7.1
purl==1.3
pycountry==17.1.8
pycparser==2.17
Pyphen==0.9.4
pysolr==3.6.0
python-dateutil==2.6.0
pytz==2017.2
rcssmin==1.0.6
requests==2.13.0
rjsmin==1.0.12
six==1.10.0
sorl-thumbnail==12.4a1
stripe==1.53.0
tinycss==0.4
Unidecode==0.4.20
WeasyPrint==0.36
webencodings==0.5.1
